requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-SocketIO==5.2.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
Flask-Login==0.6.1